+++ CHANGELOG.md
@@ -16,5 +16,5 @@ All notable changes to this project will be documented in this file.
 ## [1.0.0] - 2026-02-18
 
 ### Added
-- Initial public release of X Bookmarks Tab extension.
+- Initial public release of Totem extension.
 - GitHub Actions release pipeline that publishes an installable extension ZIP asset.